etc/python/requirements.txt
colorama==0.4.6
mysql-connector-python==9.5.0
